cleaned up controllers
